refactor: cleanup fasta reader
